+++ CHANGES.rst
@@ -3,7 +3,7 @@ Version history
 
 This library adheres to `Semantic Versioning 2.0 <http://semver.org/>`_.
 
-**UNRELEASED**
+**1.0.0**
 
 - Fixed
   ``AttributeError: 'PatchedTracebackException' object has no attribute '__cause__'``